Checkpoint going toward more depdencies on data USB key.
Go to a single partition which will be mounted on top of
/var/lib/orthanc/db-v6. There should be a templates and reports directory
there. This will be a ntfs-3g filesystem with the label CODE_DARK.

Add DEBUGV values for the Makefile so we can easily do upgrades
while doing development (4th value).

Add a udev detection for this device so it can be hotplugged.
We use a script to call systemd-mount for mounting and
umounting.

Add a postinst script. This stops orthanc after installing
kp-report.

This is all tentative.