Add generic.json template when we have a modality, but no templates for that modality.
The generic template showed the need to add a short label
for an entry. Patient age can also use that function.
This caused us to have to add a SHORTENTRY type to
the fseq/findings bundle.

Fix spawn_editor to handle filenames that have spaces in them.
It is not illegal for patient IDs to have embedded spaces.

Build dependencies for ReportPlugin needs to include libboost-all-dev.

Delete the obsolete README file.