Add add_tight_label_entry and (commented out) add_label_menu.
The former is to get both a label and a short entry shoved into column 1.

The later is for later when we figure out a place to put left/right
and upper/lower choices.

Change CTDI/DLP into tight label entry entities.

Change MODALITY into same. Not sure it should be part
of FINDINGS- really same issue as right/left or upper/lower
as these are parameters, not findings.