Add configurator json file to /code_dark/config and integrate into code_dark_rptgen. Make Code_Dark_Configurator.py code_dark_configurator and make it installable. Give it an argument to select a non-standard config file.
Fix up CDW_Configure to use jq to get data out of configurator.json and
apply that to system config- at least for Dicom parameters right now.

Have code_dark_rptgen cache the last signing physiian so that the next
time we come up, if the signer field is not already set, use the
cached signer if they're still in the authorized list.