Add a TODO file for plans.
Tweak spawn_editor yet some more so you can run it during development
as a non-orthanc user. In order to do that successfully, btw, you
have to change the code_dark mount line as so:

 #LABEL=CODE_DARK /code_dark ntfs-3g defaults,noauto,noatime,uid=1001,hide_hid_files,umask=0,dmask=27,fmask=137,noauto 0 0
 LABEL=CODE_DARK /code_dark ntfs-3g defaults,noauto,noatime,hide_hid_files,umask=0,noauto 0 0

Finally work more on code_dark_study_tool so that lists of uuids
based upon certain criteria can be presented as a table and have
a button to the right of each row for an action item to be taken.
Still early days yet.