Add back in .desktop for the configurator and make it work for both Settings menu as well as work from ~orthanc/Desktop.
It is not entirely clear whether you have to do a hover over and
allow for it to be launched. In that case, post system setup will
manage that.

We don't seem to have complete control over which screen a desktop
icon will end up on. That's something needing some work.